feat: 이벤트 수 정 페이지 제작 (#91)
