+++ src/components/index.ts
@@ -47,6 +47,7 @@ export { default as MapHeader } from './event/map-header';
 export { default as SearchBottomSheet } from './event/search-bottom-sheet';
 export { default as SearchHeader } from './event/search-header';
 export { default as EventListWithFilter } from './event/list-with-filter';
+export { default as EventForm } from './admin/event/event-form';
 // community
 export { default as EditButton } from './community/edit-button';
 export { default as ToggleSearch } from './community/toggle-search';
