+++ src/pages/admin/authority/authorize-role/index.tsx
@@ -1,8 +1,8 @@
-import { useGetPermissions } from '@/hooks';
+// import { useGetPermissions } from '@/hooks';
 
 const AuthorizeRolePage = () => {
-  const { data: permissions } = useGetPermissions();
-  console.log(permissions);
+  // const { data: permissions } = useGetPermissions();
+  // console.log(permissions);
   return <div>롤 부여</div>;
 };
