+++ src/pages/admin/event/edit/index.tsx
@@ -1,5 +1,13 @@
+import { EventForm } from '@/components';
+import { useId, useGetEventDetail } from '@/hooks';
+
 const EventEditPage = () => {
-  return <div>EventEditPage</div>;
+  const id = useId();
+  const { data: eventDetailData } = useGetEventDetail(BigInt(id));
+
+  return (
+    <EventForm mode="update" eventDetailData={eventDetailData.success?.event} />
+  );
 };
 
 export default EventEditPage;
