+++ src/utils/dateFormatter.ts
@@ -9,10 +9,7 @@ import {
 // Date 객체를 YYYY-MM-DD 형식으로 변환
 export const formatDate = (date: Date | null) => {
   if (!date) return null;
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-  return `${year}-${month}-${day}`;
+  return format(date, 'yyyy-MM-dd');
 };
 
 // YYYY-MM-DD 형식 문자열을 MM월 DD일 (요일)-선택 형식으로 변환
